requirements.txt
numpy
scipy
pyserial
opencv-python
panda-python
pandas
pyarrow
rerun-sdk